making the menu only scroll when it’s too wide
When using the standard display mode, the menu would scroll so that the first item was selected flush to the left edge, the right one on the right and the others proportionally in between. This change mankes the menu left aligned and only enables horizontal scrolling, when the menu entries would not fit onto the screen.